# feat(repr): switch internal representation to Affine