--- a/src/cmp_pairing.rs
+++ b/src/cmp_pairing.rs
@@ -1,10 +1,10 @@
 use std::cmp::Ordering;
 
-use group::{prime::PrimeCurve, GroupEncoding};
+use group::prime::PrimeCurveAffine;
 
 /// Compares two curve elements and returns their `Ordering`.
-pub fn cmp_projective<G: PrimeCurve>(x: &G, y: &G) -> Ordering {
-    let xc = x.to_affine().to_bytes();
-    let yc = y.to_affine().to_bytes();
+pub fn cmp_affine<G: PrimeCurveAffine>(x: &G, y: &G) -> Ordering {
+    let xc = x.to_bytes();
+    let yc = y.to_bytes();
     xc.as_ref().cmp(yc.as_ref())
 }