--- a/src/lib.rs
+++ b/src/lib.rs
@@ -36,7 +36,7 @@ use rand_chacha::ChaChaRng;
 use serde::{Deserialize, Serialize};
 use zeroize::Zeroize;
 
-use crate::cmp_pairing::cmp_projective;
+use crate::cmp_pairing::cmp_affine;
 use crate::convert::{derivation_index_into_fr, fr_from_bytes, g1_from_bytes, g2_from_bytes};
 pub use crate::error::{Error, Result};
 pub use crate::into_fr::IntoFr;
@@ -45,7 +45,7 @@ use crate::secret::clear_fr;
 use crate::util::sha3_256;
 
 pub use blstrs::{
-    Bls12 as PEngine, G1Affine, G1Projective as G1, G2Affine, G2Projective as G2, Scalar as Fr,
+    Bls12 as PEngine, G1Affine as G1, G1Projective, G2Affine as G2, G2Projective, Scalar as Fr,
 };
 
 /// The size of a secret key's representation in bytes.
@@ -62,17 +62,17 @@ pub const DST: &[u8; 43] = b"BLS_SIG_BLS12381G2_XMD:SHA-256_SSWU_RO_NUL_";
 
 /// A public key.
 #[derive(Deserialize, Serialize, Copy, Clone, PartialEq, Eq)]
-pub struct PublicKey(#[serde(with = "serde_impl::projective")] G1);
+pub struct PublicKey(#[serde(with = "serde_impl::affine")] G1);
 
 impl Hash for PublicKey {
     fn hash<H: Hasher>(&self, state: &mut H) {
-        self.0.to_affine().to_compressed().as_ref().hash(state);
+        self.0.to_compressed().as_ref().hash(state);
     }
 }
 
 impl fmt::Debug for PublicKey {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let uncomp = self.0.to_affine().to_uncompressed();
+        let uncomp = self.0.to_uncompressed();
         write!(f, "PublicKey({:0.10})", HexFmt(uncomp))
     }
 }
@@ -85,7 +85,7 @@ impl PartialOrd for PublicKey {
 
 impl Ord for PublicKey {
     fn cmp(&self, other: &Self) -> Ordering {
-        cmp_projective(&self.0, &other.0)
+        cmp_affine(&self.0, &other.0)
     }
 }
 
@@ -104,15 +104,15 @@ impl From<G1> for PublicKey {
 }
 
 /// Utility to convert blsttc to blst types
-impl From<PublicKey> for G1Affine {
+impl From<PublicKey> for G1Projective {
     fn from(item: PublicKey) -> Self {
-        item.0.to_affine()
+        item.0.into()
     }
 }
 
 /// Utility to convert blst to blsttc types
-impl From<G1Affine> for PublicKey {
-    fn from(item: G1Affine) -> Self {
+impl From<G1Projective> for PublicKey {
+    fn from(item: G1Projective) -> Self {
         PublicKey(item.into())
     }
 }
@@ -125,10 +125,9 @@ impl std::cmp::PartialEq<G1> for PublicKey {
 }
 
 /// Utility to compare between blsttc and blst types
-impl std::cmp::PartialEq<G1Affine> for PublicKey {
-    fn eq(&self, other: &G1Affine) -> bool {
-        // TODO is there a way to avoid the to_affine() by doing some cheap op on other?
-        &self.0.to_affine() == other
+impl std::cmp::PartialEq<G1Projective> for PublicKey {
+    fn eq(&self, other: &G1Projective) -> bool {
+        &G1Projective::from(self.0) == other
     }
 }
 
@@ -139,10 +138,9 @@ impl PublicKey {
     }
 
     /// Returns `true` if the signature matches the element of `G2`.
-    pub fn verify_g2<H: Into<G2Affine>>(&self, sig: &Signature, hash: H) -> bool {
+    pub fn verify_g2<H: Into<G2>>(&self, sig: &Signature, hash: H) -> bool {
         !self.is_zero()
-            && PEngine::pairing(&self.0.to_affine(), &hash.into())
-                == PEngine::pairing(&G1Affine::generator(), &sig.0.to_affine())
+            && PEngine::pairing(&self.0, &hash.into()) == PEngine::pairing(&G1::generator(), &sig.0)
     }
 
     /// Returns `true` if the signature matches the message.
@@ -161,12 +159,12 @@ impl PublicKey {
     /// Encrypts the message.
     pub fn encrypt_with_rng<R: RngCore, M: AsRef<[u8]>>(&self, rng: &mut R, msg: M) -> Ciphertext {
         let r: Fr = Fr::random(rng);
-        let u = G1Affine::generator().mul(r);
+        let u = G1::generator().mul(r).to_affine();
         let v: Vec<u8> = {
-            let g = self.0.to_affine().mul(r);
-            xor_with_hash(g, msg.as_ref())
+            let g = self.0.mul(r);
+            xor_with_hash(g.to_affine(), msg.as_ref())
         };
-        let w = hash_g1_g2(u, &v).to_affine().mul(r);
+        let w = hash_g1_g2(u, &v).mul(r).to_affine();
         Ciphertext(u, v, w)
     }
 
@@ -208,14 +206,14 @@ pub struct PublicKeyShare(PublicKey);
 
 impl fmt::Debug for PublicKeyShare {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let uncomp = self.0 .0.to_affine().to_uncompressed();
+        let uncomp = self.0 .0.to_uncompressed();
         write!(f, "PublicKeyShare({:0.10})", HexFmt(uncomp))
     }
 }
 
 impl PublicKeyShare {
     /// Returns `true` if the signature matches the element of `G2`.
-    pub fn verify_g2<H: Into<G2Affine>>(&self, sig: &SignatureShare, hash: H) -> bool {
+    pub fn verify_g2<H: Into<G2>>(&self, sig: &SignatureShare, hash: H) -> bool {
         self.0.verify_g2(&sig.0, hash)
     }
 
@@ -228,8 +226,7 @@ impl PublicKeyShare {
     pub fn verify_decryption_share(&self, share: &DecryptionShare, ct: &Ciphertext) -> bool {
         let Ciphertext(ref u, ref v, ref w) = *ct;
         let hash = hash_g1_g2(*u, v);
-        PEngine::pairing(&share.0.to_affine(), &hash.to_affine())
-            == PEngine::pairing(&(self.0).0.to_affine(), &w.to_affine())
+        PEngine::pairing(&share.0, &hash) == PEngine::pairing(&(self.0).0, w)
     }
 
     /// Derives a child public key share for a given index.
@@ -251,7 +248,7 @@ impl PublicKeyShare {
 /// A signature.
 // Note: Random signatures can be generated for testing.
 #[derive(Deserialize, Serialize, Clone, PartialEq, Eq)]
-pub struct Signature(#[serde(with = "serde_impl::projective")] G2);
+pub struct Signature(#[serde(with = "serde_impl::affine")] G2);
 
 impl PartialOrd for Signature {
     fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
@@ -261,33 +258,33 @@ impl PartialOrd for Signature {
 
 impl Ord for Signature {
     fn cmp(&self, other: &Self) -> Ordering {
-        cmp_projective(&self.0, &other.0)
+        cmp_affine(&self.0, &other.0)
     }
 }
 
 impl Distribution<Signature> for Standard {
     fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> Signature {
-        Signature(G2::random(rng))
+        Signature(G2Projective::random(rng).to_affine())
     }
 }
 
 impl fmt::Debug for Signature {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let uncomp = self.0.to_affine().to_uncompressed();
+        let uncomp = self.0.to_uncompressed();
         write!(f, "Signature({:0.10})", HexFmt(uncomp))
     }
 }
 
 impl Hash for Signature {
     fn hash<H: Hasher>(&self, state: &mut H) {
-        self.0.to_affine().to_compressed().as_ref().hash(state);
+        self.0.to_compressed().hash(state);
     }
 }
 
 impl Signature {
     /// Returns `true` if the signature contains an odd number of ones.
     pub fn parity(&self) -> bool {
-        let uncomp = self.0.to_affine().to_uncompressed();
+        let uncomp = self.0.to_uncompressed();
         let xor_bytes: u8 = uncomp.as_ref().iter().fold(0, |result, byte| result ^ byte);
         0 != xor_bytes.count_ones() % 2
     }
@@ -317,7 +314,7 @@ impl Distribution<SignatureShare> for Standard {
 
 impl fmt::Debug for SignatureShare {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let uncomp = (self.0).0.to_affine().to_uncompressed();
+        let uncomp = (self.0).0.to_uncompressed();
         write!(f, "SignatureShare({:0.10})", HexFmt(uncomp))
     }
 }
@@ -424,12 +421,12 @@ impl SecretKey {
 
     /// Returns the matching public key.
     pub fn public_key(&self) -> PublicKey {
-        PublicKey(G1Affine::generator() * self.0)
+        PublicKey((G1::generator() * self.0).to_affine())
     }
 
     /// Signs the given element of `G2`.
-    pub fn sign_g2<H: Into<G2Affine>>(&self, hash: H) -> Signature {
-        Signature(hash.into().mul(self.0))
+    pub fn sign_g2<H: Into<G2>>(&self, hash: H) -> Signature {
+        Signature(hash.into().mul(self.0).to_affine())
     }
 
     /// Signs the given message.
@@ -466,7 +463,7 @@ impl SecretKey {
             return None;
         }
         let Ciphertext(ref u, ref v, _) = *ct;
-        let g = u.to_affine().mul(self.0);
+        let g = u.mul(self.0).to_affine();
         Some(xor_with_hash(g, v))
     }
 
@@ -526,7 +523,7 @@ impl SecretKeyShare {
     }
 
     /// Signs the given element of `G2`.
-    pub fn sign_g2<H: Into<G2Affine>>(&self, hash: H) -> SignatureShare {
+    pub fn sign_g2<H: Into<G2>>(&self, hash: H) -> SignatureShare {
         SignatureShare(self.0.sign_g2(hash))
     }
 
@@ -545,7 +542,7 @@ impl SecretKeyShare {
 
     /// Returns a decryption share, without validating the ciphertext.
     pub fn decrypt_share_no_verify(&self, ct: &Ciphertext) -> DecryptionShare {
-        DecryptionShare(ct.0.to_affine() * (self.0).0)
+        DecryptionShare((ct.0 * (self.0).0).to_affine())
     }
 
     /// Generates a non-redacted debug string. This method differs from
@@ -574,17 +571,17 @@ impl SecretKeyShare {
 /// An encrypted message.
 #[derive(Deserialize, Serialize, Debug, Clone, PartialEq, Eq)]
 pub struct Ciphertext(
-    #[serde(with = "serde_impl::projective")] G1,
+    #[serde(with = "serde_impl::affine")] G1,
     Vec<u8>,
-    #[serde(with = "serde_impl::projective")] G2,
+    #[serde(with = "serde_impl::affine")] G2,
 );
 
 impl Hash for Ciphertext {
     fn hash<H: Hasher>(&self, state: &mut H) {
         let Ciphertext(ref u, ref v, ref w) = *self;
-        u.to_affine().to_compressed().as_ref().hash(state);
+        u.to_compressed().as_ref().hash(state);
         v.hash(state);
-        w.to_affine().to_compressed().as_ref().hash(state);
+        w.to_compressed().as_ref().hash(state);
     }
 }
 
@@ -598,9 +595,7 @@ impl Ord for Ciphertext {
     fn cmp(&self, other: &Self) -> Ordering {
         let Ciphertext(ref u0, ref v0, ref w0) = self;
         let Ciphertext(ref u1, ref v1, ref w1) = other;
-        cmp_projective(u0, u1)
-            .then(v0.cmp(v1))
-            .then(cmp_projective(w0, w1))
+        cmp_affine(u0, u1).then(v0.cmp(v1)).then(cmp_affine(w0, w1))
     }
 }
 
@@ -610,16 +605,15 @@ impl Ciphertext {
     pub fn verify(&self) -> bool {
         let Ciphertext(ref u, ref v, ref w) = *self;
         let hash = hash_g1_g2(*u, v);
-        PEngine::pairing(&G1Affine::generator(), &w.to_affine())
-            == PEngine::pairing(&u.to_affine(), &hash.to_affine())
+        PEngine::pairing(&G1::generator(), w) == PEngine::pairing(u, &hash)
     }
 
     /// Returns byte representation of Ciphertext
     pub fn to_bytes(&self) -> Vec<u8> {
         let Ciphertext(ref u, ref v, ref w) = *self;
         let mut result: Vec<u8> = Default::default();
-        result.extend(u.to_affine().to_compressed().as_ref());
-        result.extend(w.to_affine().to_compressed().as_ref());
+        result.extend(u.to_compressed().as_ref());
+        result.extend(w.to_compressed().as_ref());
         result.extend(v);
         result
     }
@@ -646,17 +640,17 @@ impl Ciphertext {
 
 /// A decryption share. A threshold of decryption shares can be used to decrypt a message.
 #[derive(Clone, Deserialize, Serialize, PartialEq, Eq)]
-pub struct DecryptionShare(#[serde(with = "serde_impl::projective")] G1);
+pub struct DecryptionShare(#[serde(with = "serde_impl::affine")] G1);
 
 impl Distribution<DecryptionShare> for Standard {
     fn sample<R: Rng + ?Sized>(&self, rng: &mut R) -> DecryptionShare {
-        DecryptionShare(G1::random(rng))
+        DecryptionShare(G1Projective::random(rng).to_affine())
     }
 }
 
 impl Hash for DecryptionShare {
     fn hash<H: Hasher>(&self, state: &mut H) {
-        self.0.to_affine().to_compressed().as_ref().hash(state);
+        self.0.to_compressed().as_ref().hash(state);
     }
 }
 
@@ -769,8 +763,10 @@ impl PublicKeySet {
     {
         let samples = shares
             .into_iter()
-            .map(|(i, share)| (i, (share.borrow().0).0));
-        Ok(Signature(interpolate(self.commit.degree(), samples)?))
+            .map(|(i, share)| (i, G2Projective::from((share.borrow().0).0)));
+        Ok(Signature(
+            interpolate(self.commit.degree(), samples)?.to_affine(),
+        ))
     }
 
     /// Combines the shares to decrypt the ciphertext.
@@ -779,8 +775,10 @@ impl PublicKeySet {
         I: IntoIterator<Item = (T, &'a DecryptionShare)>,
         T: IntoFr,
     {
-        let samples = shares.into_iter().map(|(i, share)| (i, &share.0));
-        let g = interpolate(self.commit.degree(), samples)?;
+        let samples = shares
+            .into_iter()
+            .map(|(i, share)| (i, G1Projective::from(share.0)));
+        let g = interpolate(self.commit.degree(), samples)?.to_affine();
         Ok(xor_with_hash(g, &ct.1))
     }
 
@@ -910,7 +908,7 @@ impl SecretKeySet {
 
 /// Returns a hash of the given message in `G2`.
 pub fn hash_g2<M: AsRef<[u8]>>(msg: M) -> G2 {
-    G2::hash_to_curve(msg.as_ref(), DST, &[])
+    G2Projective::hash_to_curve(msg.as_ref(), DST, &[]).to_affine()
 }
 
 /// Returns a hash of the group element and message, in the second group.
@@ -922,13 +920,13 @@ fn hash_g1_g2<M: AsRef<[u8]>>(g1: G1, msg: M) -> G2 {
     } else {
         msg.as_ref().to_vec()
     };
-    msg.extend(g1.to_affine().to_compressed().as_ref());
+    msg.extend(g1.to_compressed().as_ref());
     hash_g2(&msg)
 }
 
 /// Returns the bitwise xor of `bytes` with a sequence of pseudorandom bytes determined by `g1`.
 fn xor_with_hash(g1: G1, bytes: &[u8]) -> Vec<u8> {
-    let digest = sha3_256(g1.to_affine().to_compressed().as_ref());
+    let digest = sha3_256(&g1.to_compressed());
     let rng = ChaChaRng::from_seed(digest);
     let xor = |(a, b): (u8, &u8)| a ^ b;
     rng.sample_iter(&Standard).zip(bytes).map(xor).collect()
@@ -1029,10 +1027,10 @@ mod tests {
             let mut x = 0;
             for _ in 0..=deg {
                 x += rng.gen_range(1..5);
-                values.push((x - 1, comm.evaluate(x)));
+                values.push((x - 1, G1Projective::from(comm.evaluate(x))));
             }
             let actual = interpolate(deg, values).expect("wrong number of values");
-            assert_eq!(comm.evaluate(0), actual);
+            assert_eq!(comm.evaluate(0), actual.to_affine());
         }
     }
 
@@ -1182,22 +1180,22 @@ mod tests {
 
         assert_eq!(hash_g2(&msg), hash_g2(&msg));
         assert_ne!(hash_g2(&msg), hash_g2(&msg_end0));
-        assert_ne!(hash_g2(&msg_end0), hash_g2(&msg_end1));
+        assert_ne!(hash_g2(&msg_end0), hash_g2(msg_end1));
     }
 
     /// Some basic sanity checks for the `hash_g1_g2` function.
     #[test]
     fn test_hash_g1_g2() {
         let mut rng = rand::thread_rng();
-        let g0 = G1::random(&mut rng);
-        let g1 = G1::random(&mut rng);
+        let g0 = G1Projective::random(&mut rng).to_affine();
+        let g1 = G1Projective::random(&mut rng).to_affine();
         let msg: Vec<u8> = rng.sample_iter(&Standard).take(1000).collect();
         let msg_end0: Vec<u8> = msg.iter().chain(b"end0").cloned().collect();
         let msg_end1: Vec<u8> = msg.iter().chain(b"end1").cloned().collect();
 
         assert_eq!(hash_g1_g2(g0, &msg), hash_g1_g2(g0, &msg));
         assert_ne!(hash_g1_g2(g0, &msg), hash_g1_g2(g0, &msg_end0));
-        assert_ne!(hash_g1_g2(g0, &msg_end0), hash_g1_g2(g0, &msg_end1));
+        assert_ne!(hash_g1_g2(g0, &msg_end0), hash_g1_g2(g0, msg_end1));
         assert_ne!(hash_g1_g2(g0, &msg), hash_g1_g2(g1, &msg));
     }
 
@@ -1205,8 +1203,8 @@ mod tests {
     #[test]
     fn test_xor_with_hash() {
         let mut rng = rand::thread_rng();
-        let g0 = G1::random(&mut rng);
-        let g1 = G1::random(&mut rng);
+        let g0 = G1Projective::random(&mut rng).to_affine();
+        let g1 = G1Projective::random(&mut rng).to_affine();
         let xwh = xor_with_hash;
         assert_eq!(xwh(g0, &[0; 5]), xwh(g0, &[0; 5]));
         assert_ne!(xwh(g0, &[0; 5]), xwh(g1, &[0; 5]));
@@ -1316,7 +1314,7 @@ mod tests {
         // secret key gives same public key
         assert_eq!(pkbytes, pk.to_bytes());
         // signature matches test vector
-        let sig = sk.sign(&msgbytes);
+        let sig = sk.sign(msgbytes);
         assert_eq!(sigbytes, sig.to_bytes());
         // signature can be verified
         assert!(pk.verify(&sig, msgbytes));
@@ -1884,7 +1882,7 @@ mod tests {
         let msg = "Totally real news";
         let mut child_sig_shares = BTreeMap::default();
         for (i, child_key_share) in child_key_shares.iter() {
-            let child_sig_share = child_key_share.sign(&msg);
+            let child_sig_share = child_key_share.sign(msg);
             child_sig_shares.insert(i, child_sig_share);
         }
         // Combining the child shares creates a valid signature for the child
@@ -1918,8 +1916,8 @@ mod tests {
 
         // check that the attack works without the 0 check
         assert_eq!(
-            PEngine::pairing(&rogue_public_key.0.to_affine(), &hash.into()),
-            PEngine::pairing(&G1Affine::generator(), &rogue_sig.0.to_affine())
+            PEngine::pairing(&rogue_public_key.0, &hash),
+            PEngine::pairing(&G1::generator(), &rogue_sig.0)
         );
 
         // check that verify_g2 is protected against this attack