The current representation is Projective, this is quite large (144bytes for public keys). If we switch internal representation to Affine, we come down to 96bytes for public keys.

There's a small API breaking change, you sometimes need to pass by value when you used to be able pass by reference.

Everything else remains the same.

# size changes:
## Before this PR:
```
144	DecryptionShare - src/lib.rs:649:1: 649:74 (#0)
144	PublicKey - src/lib.rs:65:1: 65:68 (#0)
144	PublicKeyShare - src/lib.rs:207:1: 207:38 (#0)
288	Signature - src/lib.rs:254:1: 254:68 (#0)
288	SignatureShare - src/lib.rs:310:1: 310:42 (#0)
456	Ciphertext - src/lib.rs:576:1: 580:3 (#0)
```

## With this PR:
```
 96	DecryptionShare - src/lib.rs:643:1: 643:70 (#0)
 96	PublicKey - src/lib.rs:65:1: 65:64 (#0)
 96	PublicKeyShare - src/lib.rs:205:1: 205:38 (#0)
192	Signature - src/lib.rs:251:1: 251:64 (#0)
192	SignatureShare - src/lib.rs:307:1: 307:42 (#0)
312	Ciphertext - src/lib.rs:573:1: 577:3 (#0)
```

# perf changes:
All benches are roughly identical except for the `Commitment/addition` benches which are about 3x slower when we store commitments as Affine. I think this is due to the lack of `G1Affine::add_asign` forcing a lot of conversions there.
```
Commitment/addition/5   time:   [20.766 us 20.797 us 20.833 us]
                        change: [+370.86% +371.75% +372.66%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
Commitment/addition/10  time:   [37.463 us 37.536 us 37.611 us]
                        change: [+360.72% +362.00% +363.24%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
Commitment/addition/20  time:   [71.950 us 72.054 us 72.176 us]
                        change: [+366.31% +367.63% +368.90%] (p = 0.00 < 0.05)
                        Performance has regressed.
```